fix(api): waveformPeaks é [Float!]!, não [Int!]!
Achado na validação ponta-a-ponta da tarefa 1.3: com o 403 fora do caminho,
`/perfil/<username>` passou a responder
`Int cannot represent non-integer value: 0.4583` e `data: null`.

`compute_waveform_peaks` devolve RMS normalizado 0..1 (`list[float]`) — é o
que está gravado no banco de dev e em produção. O tipo GraphQL dizia
`[Int!]!`, e como o campo é non-null o erro de serialização sobe até a raiz e
zera a resposta inteira: qualquer operação que pedisse `waveformPeaks` de um
áudio real vinha vazia.

Passou invisível por dois motivos somados: o CSRF barrava o SSR antes de
chegar ao resolver, e o teste que cobria o campo semeava peaks inteiros
(`[10, 20, 30, 40, 50]`), valores que a extração real nunca produz. O teste
novo usa a forma real.

SDL regenerado.

+++ apps/api/types.py
@@ -489,7 +489,10 @@ class HymnType:
 @strawberry_django.type(hymn_models.HymnAudio)
 class HymnAudioType:
     id: strawberry.auto
-    waveform_peaks: list[int]
+    # `compute_waveform_peaks` devolve RMS normalizado 0..1 — float, não int.
+    # Estava tipado `list[int]` e todo áudio real derrubava a operação com
+    # "Int cannot represent non-integer value".
+    waveform_peaks: list[float]
     title: strawberry.auto
     source: strawberry.auto
     credits: strawberry.auto
